Refactored connection into Avatar component
